console/requirements.txt
schedule
requests
psycopg2-binary